new tasks

+++ 0x00-python_variable_annotations/0-add.py
@@ -2,6 +2,6 @@
 """Variable annotation introduction alx"""
 
 
-def add(a: float, b: float) -> int:
+def add(a: float, b: float) -> float:
     """ return the sum of a and b """
     return a + b
